requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
youtube-transcript-api==0.6.2
pydantic==2.8.2
python-dotenv==1.0.1
